+++ packages/core/src/plugins/TravenPlugin.js
@@ -17,6 +17,8 @@
 /**
  * Context passed to plugin lifecycle methods.
  * @typedef {Object} PluginContext
+ * @property {any} [editor] - TravenEditor instance (may not have a view yet during onRegister).
+ * @property {Object} [options] - Constructor options passed to TravenEditor.
  */
 
 export class TravenPlugin {